Extract duplicated methods into helper
